+++ Task1/TestTaskCFT/CarDetailModule/View/CarDetailView.swift
@@ -8,64 +8,83 @@
 
 import UIKit
 
-protocol CarDetailViewInput : UIViewController  {
+protocol ICarDetailViewInput : UIViewController  {
     func setInitialState()
     func setViewModel(viewModel: CarDetailViewModel)
 }
 
-protocol CarDetailViewOutput {
+protocol ICarDetailViewOutput {
     func callPopover(fromView: UIView, option: String)
-    func changeSelectedDataInView(type: pickerType, value: String)
+	func changeSelectedDataInView(type: PickerType, index: Int)
+	func saveSelectedTextFieldValue(type: PickerType, value: String)
     func viewDidLoadDone()
     func saveCarInDB()
+	func reloadData()
 }
 
-class CarDetailView: UIViewController, CarDetailViewInput, UITextFieldDelegate {
+class CarDetailView: UIViewController {
     
-    var output: CarDetailViewOutput?
+    var output: ICarDetailViewOutput?
     var viewModel: CarDetailViewModel?
     
     @IBOutlet weak var carModel: UITextField!
     @IBOutlet weak var carYear: UITextField!
     @IBOutlet weak var carCountry: UITextField!
     @IBOutlet weak var carBodyStyle: UITextField!
-    
+	@IBOutlet weak var carNumberLabel: UITextField!
+
     override func viewDidLoad() {
         super.viewDidLoad()
         output?.viewDidLoadDone()
     }
     
     @objc func saveCar(_ sender: Any) {
-        output!.saveCarInDB()
-    }
-    
-    func setInitialState() {
-       let rightButton = UIBarButtonItem.init(barButtonSystemItem: .save, target: self, action: #selector(saveCar(_:)))
-       self.navigationItem.rightBarButtonItem = rightButton
-       carModel.delegate = self
-       carCountry.delegate = self
-       carBodyStyle.delegate = self
-       carYear.delegate = self
-    }
-   
-    func setViewModel(viewModel: CarDetailViewModel) {
-        self.viewModel = viewModel
-        self.configureDetailView(withObject: viewModel)
-    }
-       
-    func textFieldShouldBeginEditing(_ textField: UITextField) -> Bool {
-        if output != nil  {
-            output?.callPopover(fromView: textField, option: textField.text!)
-            return false
-        }
-        return true
-    }
-    
-    private func configureDetailView(withObject object: CarDetailViewModel) {
-        self.carModel!.text = object.carModel
-        self.carYear!.text = object.carYear
-        self.carCountry!.text = object.carCountry
-        self.carBodyStyle!.text = object.carBodyStyle
+		output?.saveSelectedTextFieldValue(type: .carNumber, value: carNumberLabel.text!)
+		if ((Int(carYear.text!)) != nil) ||
+			(carYear.text!.trimmingCharacters(in: .whitespaces).isEmpty)  {
+			output?.saveSelectedTextFieldValue(type: .carYear, value: carYear.text!)
+		}
+		output?.reloadData()
+        output?.saveCarInDB()
     }
 }
 
+extension CarDetailView: UITextFieldDelegate {
+	func textFieldShouldBeginEditing(_ textField: UITextField) -> Bool {
+		if output != nil && textField.tag != 3 && textField.tag != 6 {
+			output?.callPopover(fromView: textField, option: textField.text!)
+			return false
+		}
+		return true
+	}
+}
+
+extension CarDetailView: ICarDetailViewInput
+{
+	func setInitialState() {
+	   let rightButton = UIBarButtonItem.init(barButtonSystemItem: .save, target: self, action: #selector(saveCar(_:)))
+	   self.navigationItem.rightBarButtonItem = rightButton
+	   carModel.delegate = self
+	   carCountry.delegate = self
+	   carBodyStyle.delegate = self
+	   carYear.delegate = self
+		carNumberLabel.delegate = self
+
+	}
+
+	func setViewModel(viewModel: CarDetailViewModel) {
+		self.viewModel = viewModel
+		self.configureDetailView(withObject: viewModel)
+	}
+}
+
+private extension CarDetailView
+{
+	private func configureDetailView(withObject object: CarDetailViewModel) {
+		self.carModel!.text = object.carModel
+		self.carCountry!.text = object.carCountry
+		self.carBodyStyle!.text = object.carBodyStyle
+		self.carNumberLabel.text = object.carNumber
+		self.carYear.text = object.carYear
+	}
+}
