+++ Task1/TestTaskCFT/CarDetailModule/View/PickerView.swift
@@ -7,53 +7,59 @@
 //
 
 import UIKit
-enum pickerType: String {
+enum PickerType: String {
     case carModel
     case carCountry
     case carBodyStyle
     case carYear
+	case carNumber
 }
 
-class PickerView: UIViewController, UIPickerViewDelegate,UIPickerViewDataSource {
+class PickerView: UIViewController {
     
     @IBOutlet weak var picker: UIPickerView!
     @IBOutlet weak var button: UIButton!
-    var type: pickerType?
+    var type: PickerType?
     var carService: CarService?
     var options: [String]?
     var carID: Int?
     var currentOption: String?
     var selectedIndexInPicker: Int?
-    var output: CarDetailViewOutput?
-    
+    var output: ICarDetailViewOutput?
+	var outputList: ICarsListViewOutput?
     private func selectedOption() -> [String] {
+		var itemArray = [String]()
         switch type {
         case .carModel:
-            return CarService.carModels
+			for item in CarService.CarModels.allCases {
+				itemArray.append(item.rawValue)
+			}
         case .carBodyStyle:
-            return CarService.carBodyStyle
+			for item in CarService.CarBodyStyle.allCases {
+				itemArray.append(item.rawValue)
+			}
         case .carCountry:
-            return CarService.carCountry
-        case .carYear:
-            var stringArray = [String]()
-            for value in CarService.carYear {
-                stringArray.append(String(value))
-            }
-            return stringArray
+			for item in CarService.CarCountry.allCases {
+				itemArray.append(item.rawValue)
+			}
         default:
             return [String]()
         }
+		return itemArray
     }
-    
+
     override func viewDidLoad() {
         super.viewDidLoad()
         
         picker.delegate = self
         options = selectedOption()
+		if outputList != nil {
+			options?.insert("All bodystyles", at: 0)
+		}
         button?.addTarget(self, action: #selector(saveData(_:)), for: .touchUpInside)
-        picker?.selectRow(1, inComponent: 0, animated: true)
-        
-        selectedIndexInPicker = 0
+        picker?.selectRow(0, inComponent: 0, animated: true)
+
+		var selectedIndexInPicker = 0
         if currentOption != nil {
             for i in 0 ..< options!.count {
                 if currentOption == options![i] {
@@ -61,35 +67,55 @@ class PickerView: UIViewController, UIPickerViewDelegate,UIPickerViewDataSource 
                 }
             }
         }
-        
-        picker?.selectRow(selectedIndexInPicker!, inComponent: 0, animated: true)
+		if output != nil {
+			picker?.selectRow(selectedIndexInPicker, inComponent: 0, animated: true)
+		}
+		else {
+			if selectedIndexInPicker > 0 {
+				picker?.selectRow(selectedIndexInPicker + 1, inComponent: 0, animated: true)
+			}
+		}
     }
-    
-    func pickerView(_ pickerView: UIPickerView, viewForRow row: Int, forComponent component: Int, reusing view: UIView?) -> UIView {
-        var label = view as? UILabel
-        if label == nil {
-            label = UILabel()
-        }
-        
-        label?.text = options![row]
-        label?.adjustsFontSizeToFitWidth = true
-        label?.textAlignment = .center
-        return label!
-    }
-    
+
     @objc func saveData(_ but: UIButton) {
-        let row = picker?.selectedRow(inComponent: 0)
-        output!.changeSelectedDataInView(type: type!, value: options![row!])
+		if let out = output {
+			let row = picker?.selectedRow(inComponent: 0)
+			out.changeSelectedDataInView(type: type!, index: row!)
+		}
+		if let out = outputList {
+			let row = picker?.selectedRow(inComponent: 0)
+			if row! > 0 {
+				out.filterCars(bodyStyle: CarService.CarBodyStyle(rawValue: options![row!]))
+			}
+			else {
+				out.filterCars(bodyStyle: nil)
+			}
+		}
         self.dismiss(animated: true, completion: nil)
     }
-    
-    func pickerView(_ pickerView: UIPickerView, numberOfRowsInComponent component: Int) -> Int {
-        return options!.count
-    }
-    
-    func numberOfComponents(in pickerView: UIPickerView) -> Int {
-        return 1
-    }
-    
 }
 
+extension PickerView: UIPickerViewDelegate
+{
+	func pickerView(_ pickerView: UIPickerView, viewForRow row: Int, forComponent component: Int, reusing view: UIView?) -> UIView {
+		var label = view as? UILabel
+		if label == nil {
+			label = UILabel()
+		}
+		label?.text = options![row]
+		label?.adjustsFontSizeToFitWidth = true
+		label?.textAlignment = .center
+		return label!
+	}
+}
+
+extension PickerView: UIPickerViewDataSource
+{
+	func pickerView(_ pickerView: UIPickerView, numberOfRowsInComponent component: Int) -> Int {
+		return options!.count
+	}
+
+	func numberOfComponents(in pickerView: UIPickerView) -> Int {
+		return 1
+	}
+}
