+++ Task1/TestTaskCFT/CarsListModule/View/Cell/CarsElementCell.swift
@@ -9,21 +9,36 @@
 import UIKit
 
 class CarsElementCell: UITableViewCell {
-    
+	var viewModel: CarsElementViewModel?
+	
     @IBOutlet weak var carModel: UILabel!
     @IBOutlet weak var carYear: UILabel!
     @IBOutlet weak var carCountry: UILabel!
     @IBOutlet weak var carBodyStyle: UILabel!
-    var viewModel: CarsElementViewModel?
+	@IBOutlet weak var carNumber: UILabel!
+
     
     static let reuseIdentifier = "CarsListCellIdentifier"
     
     func configureCell(withObject object: CarsElementViewModel) {
         viewModel = object
         self.carModel!.text = "Model: " + object.carModel!
-        self.carYear!.text = "Year: " + object.carYear!
         self.carBodyStyle!.text = "Body Style: " + object.carBodyStyle!
         self.carCountry!.text = "Country: " + object.carCountry!
+
+		if let year = object.carYear {
+			self.carYear!.text = "Year: " + year
+		}
+		else {
+			self.carYear!.text = "Year: -"
+		}
+		if let number = object.carNumber {
+			//self.carNumber!.isHidden = false
+			self.carNumber!.text = "Car number: " + number
+		}
+		else {
+			//self.carNumber!.isHidden = false
+		}
     }
     
     override func prepareForReuse() {
@@ -31,5 +46,6 @@ class CarsElementCell: UITableViewCell {
         self.carYear!.text = nil
         self.carBodyStyle!.text = nil
         self.carCountry!.text = nil
+		self.carNumber!.text = nil
     }
 }
