+++ Task1/TestTaskCFT/Services/Data/CarService.swift
@@ -8,86 +8,104 @@
 
 import UIKit
 
-protocol CarsListServiceInterface {
+protocol ICarsListServiceInterface {
     func getCars() -> [CarModel]?
     func deleteCar(key: String)
 }
 
-protocol DetailsCarServiceInterface {
+protocol IDetailsCarServiceInterface {
     func getCar(key: String) -> CarModel?
     func addCar(car: CarModel)
     func updateCar(car: CarModel)
 }
 
-class CarService: CarsListServiceInterface, DetailsCarServiceInterface {
-    
-    static var carModels =
-        ["Toyota", "Mazda", "Nissan", "Wolkswagen", "BMW"]
-    static var carCountry = ["Germany", "Japan", "Russia", "USA"]
-    static var carBodyStyle = ["Sedan","Universal","Coupe","Hatchback"]
-    static var carYear:[Int] = Array(2000...2014)
-    
-    func getCars() -> [CarModel]? {
-        let allValues: Data? = UserDefaults.standard.data(forKey: "cars")
-        if let cars = allValues {
-            let decodedCars = try! JSONDecoder().decode([CarModel].self, from: cars)
-            return decodedCars
-        }
-        else {
-            return [CarModel]()
-        }
-    }
-    
-    func updateCar(car: CarModel) {
-        var cars = getCars()
-        for index in 0 ..< cars!.count {
-            if cars![index].carKey == car.carKey {
-                cars![index] = car
-                break
-            }
-        }
-        setCars(cars: cars!)
-    }
-    
-    private func setCars(cars: [CarModel]) {
-        let encodedData =  try! JSONEncoder().encode(cars)
-        UserDefaults.standard.set(encodedData, forKey: "cars")
-        UserDefaults.standard.synchronize()
-    }
-    
-    func addCar(car: CarModel) {
-        let cars = getCars()
-                
-        if var carArray = cars {
-            car.carKey = UUID().uuidString
-            carArray.append(car);
-            setCars(cars: carArray)
-        }
+class CarService {
+	enum CarBodyStyle: String, CaseIterable, Codable   {
+		case Sedan
+		case Universal
+		case Coupe
+		case Hatchback
+	}
 
-    }
-    
-    func deleteCar(key: String) {
-        var cars = getCars()
-        for index in 0 ..< cars!.count {
-            if cars![index].carKey == key {
-                cars!.remove(at: index)
-                break
-            }
-        }
-        setCars(cars: cars!)
+	enum CarCountry: String, CaseIterable, Codable  {
+		case Germany
+		case Japan
+		case Russia
+		case USA
+	}
 
-    }
-    
-    func getCar(key: String) -> CarModel? {
-        let cars = getCars()
-        
-        for car in cars! {
-            if car.carKey == key {
-                return car
-            }
-        }
-       
-        return nil
-    }
+	enum CarModels: String, CaseIterable, Codable   {
+		case Toyota
+		case Mazda
+		case Nissan
+		case Wolkswagen
+		case BMW
+	}
 }
 
+extension CarService: ICarsListServiceInterface
+{
+	func getCars() -> [CarModel]? {
+		let allValues: Data? = UserDefaults.standard.data(forKey: "cars")
+		if let cars = allValues {
+			let decodedCars = try! JSONDecoder().decode([CarModel].self, from: cars)
+			return decodedCars
+		}
+		else {
+			return [CarModel]()
+		}
+	}
+
+	func deleteCar(key: String) {
+		var cars = getCars()
+		for index in 0 ..< cars!.count {
+			if cars![index].carKey == key {
+				cars!.remove(at: index)
+				break
+			}
+		}
+		setCars(cars: cars!)
+	}
+}
+
+extension CarService: IDetailsCarServiceInterface
+{
+	func updateCar(car: CarModel) {
+		var cars = getCars()
+		for index in 0 ..< cars!.count {
+			if cars![index].carKey == car.carKey {
+				cars![index] = car
+				break
+			}
+		}
+		setCars(cars: cars!)
+	}
+
+	func addCar(car: CarModel) {
+		let cars = getCars()
+		if var carArray = cars {
+			var carCopy = car
+			carCopy.carKey = UUID().uuidString
+			carArray.append(carCopy);
+			setCars(cars: carArray)
+		}
+	}
+
+	func getCar(key: String) -> CarModel? {
+		let cars = getCars()
+		for car in cars! {
+			if car.carKey == key {
+				return car
+			}
+		}
+		return nil
+	}
+}
+
+private extension CarService {
+	private func setCars(cars: [CarModel]) {
+		let encodedData =  try! JSONEncoder().encode(cars)
+		UserDefaults.standard.set(encodedData, forKey: "cars")
+		UserDefaults.standard.synchronize()
+	}
+}
