+++ Task1/TestTaskCFT/Services/Data/CarsDataService/Models/CarModel.swift
@@ -8,20 +8,11 @@
 
 import UIKit
 
-class CarModel: Codable {
-    
-    var carModel: String?
-    var carCountry: String?
-    var carYear: String?
-    var carBodyStyle: String?
+struct CarModel: Codable {
+	var model: CarService.CarModels
+	var manufacturer: CarService.CarCountry
+    var yearOfIssue: Int?
+	var body: CarService.CarBodyStyle
+	var carNumber: String?
     var carKey: String?
-    
-    init(carModel: String, carCountry: String, carYear:String,
-         carBodyStyle: String, key: String?) {
-        self.carModel = carModel
-        self.carCountry = carCountry
-        self.carYear = carYear
-        self.carBodyStyle = carBodyStyle
-        self.carKey = key
-    }
 }
