--- TestTaskCFT.xcodeproj/xcuserdata/andreybelkin.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Bucket
-   uuid = "C05B7081-8964-474A-9AE1-A279949ADBB2"
-   type = "1"
-   version = "2.0">
-</Bucket>
